Agrega opcion para canejar puntos
