+++ package.json
@@ -15,7 +15,7 @@
     "@heroicons/react": "^2.0.18",
     "@hookform/resolvers": "^3.1.1",
     "@next-auth/prisma-adapter": "^1.0.7",
-    "@prisma/client": "^5.0.0",
+    "@prisma/client": "^5.1.1",
     "@radix-ui/react-dialog": "^1.0.4",
     "@radix-ui/react-progress": "^1.0.3",
     "@trpc/client": "^9.27.4",
@@ -25,7 +25,7 @@
     "clsx": "^2.0.0",
     "date-fns": "^2.30.0",
     "discord-webhook-node": "^1.1.8",
-    "framer-motion": "^10.13.1",
+    "framer-motion": "^10.15.0",
     "next": "13.4.12",
     "next-auth": "^4.22.3",
     "next-safe-action": "^3.0.1",
@@ -42,21 +42,21 @@
     "tailwind-merge": "^1.14.0",
     "tailwindcss-radix": "^2.8.0",
     "zod": "^3.21.4",
-    "zustand": "^4.3.9"
+    "zustand": "^4.4.0"
   },
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^4.1.0",
     "@tailwindcss/forms": "^0.5.4",
-    "@types/node": "20.4.5",
-    "@types/react": "18.2.17",
+    "@types/node": "20.4.7",
+    "@types/react": "18.2.18",
     "@types/react-dom": "18.2.7",
     "autoprefixer": "^10.4.14",
-    "eslint": "8.45.0",
+    "eslint": "8.46.0",
     "eslint-config-next": "13.4.12",
     "postcss": "^8.4.27",
-    "prettier": "^3.0.0",
+    "prettier": "^3.0.1",
     "prettier-plugin-tailwindcss": "^0.4.1",
-    "prisma": "^5.0.0",
+    "prisma": "^5.1.1",
     "tailwindcss": "^3.3.3",
     "typescript": "5.1.6"
   }
